requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.7
djangorestframework==3.15.2
djoser==2.3.1
firebase_admin==6.7.0
Requests==2.32.3
daphne
django-cors-headers
psycopg2
django-extensions
channels_redis